backend/requirements.txt
flask
google-generativeai
qdrant-client
langchain
pypdf
flask-sqlalchemy         
flask-migrate             
psycopg2-binary          
flask-bcrypt              
flask-jwt-extended       
python-dotenv
flask-marshmallow
marshmallow-sqlalchemy
langchain